requirements.txt
django==4.1.5
django-allauth==0.52.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
whitenoise==6.3.0
unidecode==1.3.6
weasyprint==57.2
fcm-django==1.0.12
gunicorn==20.1.0
django-oauth-toolkit==2.2.0
django-celery-beat==2.4.0
pwned-passwords-django==1.6.1
pre-commit==3.0.1
celery~=5.2.7
